improved floor entity management

+++ src/Xabbo.Core/Room/HeightmapTile.cs
@@ -11,7 +11,7 @@ public class HeightmapTile : IHeightmapTile, IComposer
     public bool IsFloor { get; set; }
     public bool IsBlocked { get; set; }
     public bool IsFree => IsFloor && !IsBlocked;
-    public double Height { get; set; }
+    public float Height { get; set; }
 
     /// <summary>
     /// Constructs a new heightmap tile with the specified coordinates and encoded value.
@@ -35,7 +35,7 @@ public class HeightmapTile : IHeightmapTile, IComposer
     {
         IsFloor = value >= 0;
         IsBlocked = (value & 0x4000) != 0;
-        Height = value >= 0 ? ((value & 0x3FFF) / 256.0) : -1;
+        Height = value >= 0 ? ((value & 0x3FFF) / 256f) : -1;
     }
 
     void IComposer.Compose(in PacketWriter p)
@@ -43,7 +43,7 @@ public class HeightmapTile : IHeightmapTile, IComposer
         p.WriteShort((short)(
             (IsFloor ? 0x0000 : 0x8000) |
             (IsBlocked ? 0x4000 : 0x0000) |
-            ((int)(Height * 256.0) & 0x3FFF)
+            ((int)(Height * 256) & 0x3FFF)
         ));
     }
 }
